Move test_default_with_assignee_and_status into test_steps
